@@ -17,13 +17,13 @@ Gem::Specification.new do |s|
   s.files = Dir["{app,config,db,lib}/**/*", "MIT-LICENSE", "Rakefile", "README.rdoc"]
   s.test_files = Dir["spec/**/*"]
 
-  s.add_dependency "rails", "~> 4.1.8"
+  s.add_dependency "rails", "~> 5.2.2"
   s.add_dependency "active_model_serializers", "~> 0.8.0"
   s.add_dependency "ims-lti"
-  s.add_dependency "rest-client"
+  s.add_dependency "rest-client", "2.0.2"
   s.add_dependency "curb"
-  s.add_dependency "mimemagic"
+  s.add_dependency "mimemagic", "0.3.2"
   s.add_dependency "jwt"
 
-  s.add_development_dependency "sqlite3"
+  s.add_development_dependency "sqlite3", "1.3.13"
 end
