@@ -1,6 +1,6 @@
 module Skydrive
   class UsersController < ApplicationController
-    before_filter :ensure_authenticated_user, except: [:create]
+    before_action :ensure_authenticated_user, except: [:create]
 
     # Returns list of users. This requires authorization
     def index
