@@ -1,4 +1,4 @@
-class CreateApiKeys < ActiveRecord::Migration
+class CreateApiKeys < ActiveRecord::Migration[4.2]
   def change
     create_table :api_keys do |t|
       t.integer :user_id
