@@ -1,4 +1,4 @@
-class AddPersonRequestInfoToAccounts < ActiveRecord::Migration
+class AddPersonRequestInfoToAccounts < ActiveRecord::Migration[4.2]
   def change
     add_column :skydrive_accounts, :name, :string
     add_column :skydrive_accounts, :email, :string
