@@ -15,7 +15,7 @@ module Skydrive
     isolate_namespace Skydrive
 
     initializer :append_migrations do |app|
-      unless app.root.to_s == root.to_s
+      unless app.root.to_s.include?(root.to_s)
         config.paths["db/migrate"].expanded.each do |expanded_path|
           app.config.paths["db/migrate"] << expanded_path
         end
